editable user profile + simplest way to make roles + post seeder(with tags and categories) + homepage view (using a ready to use them)

+++ app/Http/Controllers/CategoriesController.php
@@ -110,6 +110,9 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0){
+            return redirect()->back()->with('error', 'Category cannot be deleted because it has some posts');
+        }
         $category->delete();
 
         return redirect(route('categories.index'))->with('success', 'Category Deleted Successfully');
